+++ packages/react-ui/src/widgets/BarWidgetUI/BarSkeleton.js
@@ -1,14 +1,14 @@
 import { Box, Skeleton } from '@mui/material';
-import { SkeletonBarsGrid, SkeletonBarItem } from '../SkeletonWidgets';
+import { SkeletonBarsGrid, SkeletonBarItem, SKELETON_HEIGHT } from '../SkeletonWidgets';
 
-const BarWidgetLoadingUI = ({ height }) => {
+const BarSkeleton = ({ height }) => {
   return (
     <>
       <Box mb={2}>
         <Skeleton width={48} height={8} />
       </Box>
 
-      <SkeletonBarsGrid style={{ height: height }}>
+      <SkeletonBarsGrid style={{ height: height || SKELETON_HEIGHT }}>
         <SkeletonBarItem variant='rectangular' height='20%' />
         <SkeletonBarItem variant='rectangular' height='40%' />
         <SkeletonBarItem variant='rectangular' height='60%' />
@@ -19,4 +19,4 @@ const BarWidgetLoadingUI = ({ height }) => {
   );
 };
 
-export default BarWidgetLoadingUI;
+export default BarSkeleton;